requirements.txt
flask==2.3.3 
flask-restful==0.3.10 
flask-sqlalchemy==3.0.5 
flask-migrate==4.0.4 
flask-jwt-extended==4.5.2 
marshmallow==3.19.0 
pytest==7.4.0 
python-dotenv==1.0.0
requests==2.31.0
flask-oauthlib==0.9.6
